Safari-611.1.10/LayoutTests/fast/css/CSSStyleDeclaration-property-setter-expected.txt

Test that setting CSS properties via CSSStyleDeclaration behaves as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.style.cssText is ""
div.style.backgroundColor = 'red'
PASS div.style.cssText is "background-color: red;"
PASS div.style.getPropertyValue('background-color') is "red"
PASS div.style.length is 1

div.style.backgroundColor = ''
PASS div.style.cssText is ""
PASS div.style.getPropertyValue('background-color') is ""
PASS div.style.length is 0

div.style.backgroundColor = 'red'
PASS div.style.cssText is "background-color: red;"
PASS div.style.getPropertyValue('background-color') is "red"
PASS div.style.length is 1

div.style.backgroundColor = null
PASS div.style.cssText is ""
PASS div.style.getPropertyValue('background-color') is ""
PASS div.style.length is 0
PASS successfullyParsed is true

TEST COMPLETE
